Re-factor upload helper
# Conflicts:
#	js/ui/file_uploader.js
#	js/ui/file_uploader_helper.js
